receipt_reader/requirements.txt
numpy==1.24.4
packaging==24.0
pandas==2.0.3
pillow==10.3.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
